backend/requirements.txt
django>=5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
psycopg2-binary>=2.9
Pillow>=10.0
python-decouple>=3.8
dj-database-url>=2.1
django-filter>=23.5
requests>=2.31
